% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Beta2.R
\name{cdf_B2}
\alias{cdf_B2}
\title{cdf of the Beta Prime (Beta-2) Distribution}
\usage{
cdf_B2(y, b, p, q)
}
\arguments{
\item{y}{Numeric vector of evaluation points.}

\item{b}{Scale parameter \eqn{b > 0}.}

\item{p}{First shape parameter \eqn{p > 0}.}

\item{q}{Second shape parameter \eqn{q > 0}.}
}
\value{
Numeric vector of cdf values.
}
\description{
Computes the cumulative distribution function (cdf) of a **Beta Prime
(Beta-2)** distribution:

\deqn{F(y \mid p, q, b) = I_{\,y/(b+y)}(p, q)}

where \eqn{I_z(p,q)} is the regularized incomplete beta function.
}
\examples{
cdf_B2(y = 1:5, b = 2, p = 3, q = 4)

}
