% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lognormal.R
\name{cdf_LN}
\alias{cdf_LN}
\title{Lognormal cdf Expressed Using (mean, sigma)}
\usage{
cdf_LN(y, mean_y, sigma)
}
\arguments{
\item{y}{numerical vector for the mean income}

\item{mean_y}{numerical value, average income}

\item{sigma}{numerical value, lognormal scale positive parameter}
}
\value{
cdf, probability between 0 and 1
}
\description{
Lognormal cdf Expressed Using (mean, sigma)
}
\examples{
cdf_LN(1500, mean_y = 6000, 0.6)
[1] 0.02218964
cdf_LN(c(1500,7500), mean_y = 6000, 0.6)
[1] 0.02218964 0.74917820
}
