% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewPareto.R
\name{CDF_NP}
\alias{CDF_NP}
\title{New Pareto CDF}
\usage{
CDF_NP(y, shape, scale)
}
\arguments{
\item{y}{Quantile/income.}

\item{shape}{Shape parameter.}

\item{scale}{Scale parameter.}
}
\value{
CDF value.
}
\description{
New Pareto CDF
}
\examples{
CDF_NP(2000, shape= 2, scale=100)
#[1] 0.9950125
}
