% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layer I - estimate parameter.R
\name{estimate_param}
\alias{estimate_param}
\title{Fast Estimation of Parameters Using Mean and Gini}
\usage{
estimate_param(
  dist = c("LN", "NP", "FISK"),
  mean_y,
  Gini,
  se_mean = NA,
  se_Gini = NA
)
}
\arguments{
\item{dist}{character vector, one of "LN", "NP", "FISK"}

\item{mean_y}{Mean of the data (must be > 0)}

\item{Gini}{Gini coefficient in (0,1)}

\item{se_mean}{Standard error of mean (optional)}

\item{se_Gini}{Standard error of Gini (optional)}
}
\value{
A list with:
\describe{
\item{par}{c(shape, scale)}
\item{se}{c(se_shape, se_scale)}
}
}
\description{
Computes:
distribution parameters as in \code{\link{fit_Fisk_param}},
or \code{\link{fit_LN_param}} or \code{\link{fit_NP_param}}.
}
\details{
Optionally computes delta-method standard errors.
}
\examples{
res = estimate_param(dist = "NP", mean_y = 6000, Gini = 0.36, se_mean = 250, se_Gini = 0.01)
# res$NP
# $par
# shape       scale
# 2.128961 2523.457703
#
# $se
# se_shape se_scale
# 0.04810901 125.06197149


res = estimate_param(dist = c("LN","NP"), mean_y = 6000, Gini = 0.36, se_mean = 250, se_Gini = 0.01)
# $FISK
# [1] NA
#
# $LN
# $LN$par
# mu    sigma
# 8.480773 0.661426
#
# $LN$se
# se_mu       se_sigma
# 0.04367100 0.01977307
#
#
# $NP
# $NP$par
# shape       scale
# 2.128961 2523.457703
#
# $NP$se
# se_shape se_scale
# 0.04810901 125.06197149
}
