% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FISK.R
\name{fit_FISK_param}
\alias{fit_FISK_param}
\title{Fast Parameter Estimation for Fisk Distribution Using Only (Mean, Gini)}
\usage{
fit_FISK_param(mean_y, Gini, se_mean = NA, se_Gini = NA)
}
\arguments{
\item{mean_y}{numerical value, observed mean of the variable.}

\item{Gini}{Numeric in (0,1). The Gini coefficient}

\item{se_mean}{Standard error of mean (optional).}

\item{se_Gini}{Standard error of Gini (optional).}
}
\value{
A list with:
  \describe{
    \item{par}{Named vector: c(shape, scale)}
    \item{se}{Named vector: c(se_shape, se_scale)}
  }
}
\description{
Computes the Fisk (log-logistic) parameters using moment-matching:
  shape = 1/Gini
  scale = mean / [Gamma(1+1/shape) * Gamma(1-1/shape)]
}
\details{
Optionally propagates standard errors using the delta method.
}
\examples{
fit <- fit_FISK_param(mean_y= 20000, Gini = 0.32)
fit
$par
shape     scale
3.125 16797.370

$se
se_shape se_scale
NA       NA
}
