% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lognormal.R
\name{fit_LN_param}
\alias{fit_LN_param}
\title{Fast Estimation of Lognormal Parameters Using Mean and Gini}
\usage{
fit_LN_param(mean_y, Gini, se_mean = NA, se_Gini = NA)
}
\arguments{
\item{mean_y}{Mean of the data (must be > 0)}

\item{Gini}{Gini coefficient in (0,1)}

\item{se_mean}{Standard error of mean (optional)}

\item{se_Gini}{Standard error of Gini (optional)}
}
\value{
A list with:
\describe{
\item{par}{c(mu, sigma)}
\item{se}{c(se_mu, se_sigma)}
}
}
\description{
Computes:
sigma = sqrt(2) * qnorm((1+Gini)/2)
mu    = log(mean) - sigma^2/2
}
\details{
Optionally computes delta-method standard errors.
}
\examples{
fit_LN_param(6000, 0.36, se_mean = 250, se_Gini = 0.01)
$par
# mu    sigma
# 8.480773 0.661426
#
# $se
# se_mu   se_sigma
# 0.04367100 0.01977307
}
