% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewPareto.R
\name{fit_NP_param}
\alias{fit_NP_param}
\title{Fast Estimation of NewPareto Parameters Using Mean and Gini}
\usage{
fit_NP_param(mean_y, Gini, se_mean = NA, se_Gini = NA)
}
\arguments{
\item{mean_y}{Mean of the data (must be > 0)}

\item{Gini}{Gini coefficient in (0,1)}

\item{se_mean}{Standard error of mean (optional)}

\item{se_Gini}{Standard error of Gini (optional)}
}
\value{
A list with:
\describe{
\item{par}{c(shape, scale)}
\item{se}{c(se_shape, se_scale)}
}
}
\description{
Computes:
shape as in \code{\link{shape_NP}}
scale  as in \code{\link{scale_NP}}
}
\details{
Optionally computes delta-method standard errors as in
\code{\link{se_shape_NP}} and  \code{\link{se_scale_NP}}
}
\examples{
fit_NP_param(6000, 0.36, se_mean = 250, se_Gini = 0.01)
$par
# shape       scale
#  2.128961  2523.457703
#
# se_shape     se_scale
# 0.04810901 125.06197149
}
