% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FISK.R
\name{HC_se_FISK}
\alias{HC_se_FISK}
\title{Standard Error of Headcount Ratio Under Fisk Model}
\usage{
HC_se_FISK(y, shape, scale, se_shape, se_scale, cov_shape_scale = 0)
}
\arguments{
\item{shape}{Shape parameter.}

\item{scale}{Scale parameter.}

\item{se_shape}{SE of shape.}

\item{se_scale}{SE of scale.}

\item{cov_shape_scale}{Optional covariance term.}

\item{x}{Poverty line.}
}
\value{
Standard error of the headcount ratio.
}
\description{
Uses the delta method for:
\deqn{ HC = 1 - F(x) }
}
\examples{
HC_se_FISK(2, shape = 3, scale = 2, se_shape = 0.1, se_scale = 0.2)
}
