% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lognormal.R
\name{HC_se_LN}
\alias{HC_se_LN}
\title{Standard Error of Lognormal Headcount Ratio}
\usage{
HC_se_LN(pov_line, mean_y, sigma, se_mean, se_sigma, cov_mu_sigma = 0)
}
\arguments{
\item{mean_y}{numerical value, observed mean of the variable}

\item{sigma}{numerical value, lognormal scale positive parameter}

\item{se_mean}{standard error of the mean}

\item{se_sigma}{standard error of sigma parameter}

\item{cov_mu_sigma}{Covariance of (mu, sigma), default 0}

\item{p}{Poverty line}
}
\value{
standard error of the headcount
}
\description{
Standard Error of Lognormal Headcount Ratio
}
\examples{
HC_se_LN(900, mean_y = 2000, sigma = 0.36, se_mean = 250, se_sigma = 0.01)
0.01767621 # 1.76\%
}
