% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FISK.R
\name{Lorenz_FISK}
\alias{Lorenz_FISK}
\title{Lorenz Curve of the Fisk Distribution}
\usage{
Lorenz_FISK(p, shape)
}
\arguments{
\item{shape}{Shape parameter (> 1).}

\item{P}{Vector of cumulative population shares in (0,1).}
}
\value{
Vector of Lorenz curve values L(P).
}
\description{
Lorenz Curve of the Fisk Distribution
}
\details{
The Lorenz curve is:
\deqn{
L(p) = B(p; 1 + 1/shape,\; 1 - 1/shape)
}
where B is the incomplete beta CDF.
}
\examples{
Lorenz_FISK(p= seq(0.1, 1, by=0.1), shape = 3)
}
