% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lognormal.R
\name{Lorenz_LN}
\alias{Lorenz_LN}
\title{Lognormal Lorenz Curve}
\usage{
Lorenz_LN(p = seq(0.1, 1, 0.1), Gini)
}
\arguments{
\item{P}{Quantiles in \link{0,1}}

\item{sigma}{numerical value, lognormal scale positive parameter}
}
\value{
Lorenz curve values.
}
\description{
Computes L(p) = Φ(Φ⁻¹(p) − sigma)
}
\examples{
Lorenz_LN(Gini = 0.36)
# [1] 0.02600944 0.06641343 0.11784541 0.18015538 0.25416958 0.34160788 0.44550533 0.57150036
# [9] 0.73241244 1.00000000
}
