% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewPareto.R
\name{Lorenz_NP}
\alias{Lorenz_NP}
\title{Fast Lorenz Curve for New Pareto Distribution}
\usage{
Lorenz_NP(
  p = seq(0.1, 1, by = 0.1),
  shape,
  n_grid = 3000,
  cluster_power = 3,
  eps = 1e-12
)
}
\arguments{
\item{p}{Probabilities in \link{0,1}.}

\item{shape}{Shape parameter.}

\item{n_grid}{Grid resolution.}

\item{cluster_power}{Grid clustering exponent.}
}
\value{
Lorenz curve values.
}
\description{
Approximates:
L(p) = ∫₀ᵖ ∫₀ᵗ ((1+u)/(1-u))^(1/shape) du dt / denom
}
\examples{
Lorenz_NP(shape=3)
# [1] 0.05919209 0.12251535 0.19040316 0.26346712 0.34259232 0.42912127 0.52525375 0.63509513
# [9] 0.76851956 1.00000000
}
