% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FISK.R
\name{loss_Fisk}
\alias{loss_Fisk}
\title{Loss function for Fisk (Log-Logistic) distribution}
\usage{
loss_Fisk(
  params,
  pc.inc,
  gini.e,
  y,
  p_dec = seq(0.1, 0.9, length.out = length(y) + 1)
)
}
\arguments{
\item{params}{c(shape, scale)}

\item{pc.inc}{numerical value, average income}

\item{gini.e}{Numeric in (0,1). The Gini coefficient}

\item{y}{Observed non-cumulative income shares}

\item{p_dec}{Probabilities (e.g. seq(.1,.9,length.out=length(y)+1))}
}
\value{
Scalar loss value
}
\description{
Loss function for Fisk (Log-Logistic) distribution
}
