% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FISK.R
\name{mean_Fisk}
\alias{mean_Fisk}
\title{Mean of the Fisk Distribution}
\usage{
mean_Fisk(shape, scale)
}
\arguments{
\item{shape}{Shape parameter (> 1). Mean exists only for shape > 1.}

\item{scale}{Scale parameter (> 0).}
}
\value{
Mean value, or NA if the mean is undefined.
}
\description{
Mean of the Fisk Distribution
}
\examples{
mean_Fisk(scale = 2, shape = 3)
}
