% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lognormal.R
\name{mu_LN}
\alias{mu_LN}
\title{Lognormal Meanlog (mu)}
\usage{
mu_LN(mean_y, sigma)
}
\arguments{
\item{mean_y}{numerical value, observed mean of the variable}

\item{sigma}{numerical value, lognormal scale positive parameter}
}
\value{
mu (meanlog)
}
\description{
For a lognormal distribution:
  mean = exp(mu + sigma^2 / 2)
so
  mu = log(mean) - sigma^2/2
}
\examples{
mu_LN(1000, 0.36)
[1] 6.8429
}
