% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lognormal.R
\name{pdf_LN}
\alias{pdf_LN}
\title{Lognormal PDF (fast)}
\usage{
pdf_LN(y, mu, sigma)
}
\arguments{
\item{y}{numerical vector of the variable}

\item{mu}{numerical value, lognormal location parameter}

\item{sigma}{numerical value, lognormal scale positive parameter}
}
\value{
Density, probability between 0 and 1
}
\description{
Lognormal PDF (fast)
}
\examples{
pdf_LN(100, 6, 12)
[1] 0.0003302136
pdf_LN(c(100, 200), 6, 12)
[1] 0.0003302136 0.0001659420
}
