% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewPareto.R
\name{pdf_NP}
\alias{pdf_NP}
\title{New Pareto PDF}
\usage{
pdf_NP(y, shape, scale)
}
\arguments{
\item{shape}{Shape parameter.}

\item{scale}{Scale (minimum income).}

\item{p}{Income values.}
}
\value{
Density values.
}
\description{
New Pareto PDF
}
\examples{
pdf_NP(2000, shape= 2, scale=100)
#[1]0.000004975093
}
