% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewPareto.R
\name{Quantile_NP}
\alias{Quantile_NP}
\title{New Pareto Quantile function}
\usage{
Quantile_NP(p, shape, scale)
}
\arguments{
\item{p}{numerical vector, probability between 0 and  1.}

\item{shape}{Shape parameter.}

\item{scale}{Scale parameter.}
}
\value{
quantile value.
}
\description{
New Pareto Quantile function
}
\examples{
Quantile_NP(0.2, shape= 2, scale=100)
#[1] 122.4745
}
