% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewPareto.R
\name{random_NP}
\alias{random_NP}
\title{New Pareto Random generating function}
\usage{
random_NP(u, shape, scale)
}
\arguments{
\item{u}{numerical vector, uniform number between 0 and  1.}

\item{shape}{Shape parameter.}

\item{scale}{Scale parameter.}
}
\value{
NP value.
}
\description{
New Pareto Random generating function
}
\examples{
random_NP(0.2, shape= 2, scale=100)
#[1] 300
}
