% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FISK.R
\name{scale_FISK}
\alias{scale_FISK}
\title{Mean-Matching Scale Parameter for Fisk Distribution}
\usage{
scale_FISK(shape, mean_y)
}
\arguments{
\item{shape}{Shape parameter.}

\item{mean_y}{numerical value, observed mean of the variable.}
}
\value{
Scale parameter consistent with the specified mean.
}
\description{
Mean-Matching Scale Parameter for Fisk Distribution
}
\examples{
scale_FISK (3, 4000)
}
