% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewPareto.R
\name{scale_NP}
\alias{scale_NP}
\title{Estimate New Pareto Scale Parameter}
\usage{
scale_NP(mean_y, shape)
}
\arguments{
\item{mean_y}{Mean income.}

\item{shape}{Previously estimated shape parameter.}
}
\value{
Estimated scale parameter.
}
\description{
Solves for \eqn{\theta} from:
\deqn{E[Y] = \frac{2\alpha\theta \, \mathrm{Re}\{\,_2F_1(2,2-(1+\alpha)/\alpha;3-(1+\alpha)/\alpha;-1)\}}{\alpha - 1}}
}
\examples{
scale_NP(mean_y= 8000, shape= 0.4)
515.4424
}
