% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewPareto.R
\name{scale_NP_func_wrapper}
\alias{scale_NP_func_wrapper}
\title{Safe Wrapper for scale_NP()}
\usage{
scale_NP_func_wrapper(x)
}
\arguments{
\item{x}{Numeric vector: c(mean, shape)}
}
\value{
scalar scale or NA
}
\description{
Safe Wrapper for scale_NP()
}
