% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lognormal.R
\name{se_mu_LN}
\alias{se_mu_LN}
\title{Standard Error of Lognormal mu}
\usage{
se_mu_LN(mean_y, Gini, se_mean, se_Gini)
}
\arguments{
\item{mean_y}{numerical value, Mean of the variable.}

\item{Gini}{Numeric in (0,1). The Gini coefficient.}

\item{se_mean}{standard error of the mean}

\item{se_Gini}{standard error of the Gini index}
}
\value{
positive value, standard error of mu
}
\description{
Uses delta method based on partial derivatives wrt mean and Gini.
}
\examples{
se_mu_LN(1000,0.36, 250, 0.01)
[1] 0.2503419
}
