% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewPareto.R
\name{se_scale_NP}
\alias{se_scale_NP}
\title{Standard Error for New Pareto Scale Parameter}
\usage{
se_scale_NP(
  mean_y,
  Gini,
  se_mean,
  se_Gini,
  shape = NA,
  se_shape = NA,
  scale = NA,
  cov_mean_gini = 0,
  eps = 1e-06
)
}
\arguments{
\item{mean_y}{Mean income.}

\item{Gini}{Gini coefficient.}

\item{se_mean}{SE(mean).}

\item{se_Gini}{SE(Gini).}

\item{shape}{Optional shape (cached).}

\item{se_shape}{Optional SE(shape).}

\item{scale}{Optional scale (cached).}

\item{cov_mean_gini}{Cov(mean, Gini).}

\item{eps}{finite-difference step.}
}
\value{
Standard error of scale.
}
\description{
Standard Error for New Pareto Scale Parameter
}
\examples{
se_scale_NP(mean_y = 6000, se_mean = 250, Gini =0.36, se_Gini = 0.1, shape= 0.3, se_shape = 0.48, scale= 150)
#[1] 0.
}
