% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewPareto.R
\name{se_shape_NP}
\alias{se_shape_NP}
\title{Standard Error for New Pareto Shape}
\usage{
se_shape_NP(Gini, se_Gini, shape = NA, eps = 0.000001)
}
\arguments{
\item{Gini}{Gini coefficient.}

\item{se_Gini}{SE(Gini).}

\item{shape}{Optional precomputed shape.}

\item{eps}{Finite difference step.}
}
\value{
Standard error of shape.
}
\description{
Standard Error for New Pareto Shape
}
\examples{
se_shape_NP(Gini =0.36, se_Gini = 0.1, shape= 0.3)
#[1] 0.4810901
}
