% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lognormal.R
\name{se_sigma_LN}
\alias{se_sigma_LN}
\title{Standard Error of Lognormal Sigma From Gini}
\usage{
se_sigma_LN(mean_y, Gini, se_Gini)
}
\arguments{
\item{mean_y}{numerical value, observed mean of the variable. Not used (kept for consistency).}

\item{Gini}{Numeric in (0,1). The Gini coefficient.}

\item{se_Gini}{Standard error of the Gini index.}
}
\value{
positive value, Standard error of sigma
}
\description{
Uses delta method: SE(sigma) = |d sigma / d G| * SE(Gini)
}
\examples{
se_sigma_LN(1000,0.36, 0.01)
[1] 0.01977
}
