% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FISK.R
\name{shape_FISK}
\alias{shape_FISK}
\title{Mean-Matching Shape Parameter for Fisk Distribution}
\usage{
shape_FISK(Gini)
}
\arguments{
\item{Gini}{Numeric in (0,1). The Gini coefficient}
}
\value{
Shape parameter.
}
\description{
Mean-Matching Shape Parameter for Fisk Distribution
}
\examples{
shape_FISK (0.367)
}
