% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewPareto.R
\name{shape_NP}
\alias{shape_NP}
\title{Estimate New Pareto Shape Parameter}
\usage{
shape_NP(Gini)
}
\arguments{
\item{Gini}{The target Gini coefficient in (0,1).}
}
\value{
Estimated shape parameter \eqn{\alpha}.
}
\description{
Estimates the shape parameter \eqn{\alpha} of the New Pareto distribution
by solving:
\deqn{G(\alpha) = Gini}
}
\details{
using derivative-free nonlinear optimization (COBYLA).


The optimizer minimizes:
\deqn{|G_NP(\alpha) - Gini|}

subject to:
\deqn{\alpha \ge 1.1.}
}
\examples{
shape_NP(0.36)
#[1] 2.128961
}
