% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewPareto.R
\name{shape_NP_deriv}
\alias{shape_NP_deriv}
\title{Numerical Derivative of Shape wrt Gini}
\usage{
shape_NP_deriv(Gini)
}
\arguments{
\item{Gini}{Gini coefficient.}

\item{eps}{Step size.}
}
\value{
Numeric derivative d(alpha)/d(G).
}
\description{
Numerical Derivative of Shape wrt Gini
}
