% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewPareto.R
\name{shape_NP_func_wrapper}
\alias{shape_NP_func_wrapper}
\title{Safe Wrapper for shape_NP()}
\usage{
shape_NP_func_wrapper(Gini)
}
\arguments{
\item{Gini}{Gini coefficient.}
}
\value{
scalar numeric shape or NA.
}
\description{
Safe Wrapper for shape_NP()
}
