% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lognormal.R
\name{Sigma_LN}
\alias{Sigma_LN}
\title{Lognormal Sigma From Gini}
\usage{
Sigma_LN(Gini)
}
\arguments{
\item{Gini}{Numeric in (0,1). The Gini coefficient.}
}
\value{
The lognormal sigma.
}
\description{
Computes the lognormal sigma parameter from the Gini coefficient using:
  sigma = sqrt(2) * qnorm((1 + Gini)/2)
}
\examples{
Sigma_LN(0.36)
[1] 0.661426
}
