% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FISK.R
\name{Gini_FISK}
\alias{Gini_FISK}
\title{Gini Coefficient of the Fisk Distribution}
\usage{
Gini_FISK(shape)
}
\arguments{
\item{shape}{Shape parameter (> 0).}
}
\value{
Gini coefficient.
}
\description{
Gini Coefficient of the Fisk Distribution
}
\examples{
Gini_FISK(3)
}
