% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FISK.R
\name{scale_Fisk}
\alias{scale_Fisk}
\title{Mean-Matching Scale Parameter for Fisk Distribution}
\usage{
scale_Fisk(shape, mean_y)
}
\arguments{
\item{shape}{Shape parameter.}

\item{mean_y}{Target mean income.}
}
\value{
Scale parameter consistent with the specified mean.
}
\description{
Mean-Matching Scale Parameter for Fisk Distribution
}
\examples{
scale_Fisk (3, 4000)
}
