% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FISK.R
\name{se_scale_FISK}
\alias{se_scale_FISK}
\title{Standard Error of Fisk Scale Parameter (Delta Method)}
\usage{
se_scale_FISK(mean_y, Gini, se_mean, se_Gini, shape_FISK = NA)
}
\arguments{
\item{mean_y}{numerical value, observed mean of the variable.}

\item{Gini}{Numeric in (0,1). The Gini coefficient}

\item{se_mean}{Standard error of mean.}

\item{se_Gini}{Standard error of Gini.}

\item{shape_FISK}{Optional shape value (otherwise computed from Gini).}
}
\value{
Standard error of the Fisk scale parameter.
}
\description{
Standard Error of Fisk Scale Parameter (Delta Method)
}
\examples{
se_scale_FISK (4000,0.36, 750,0.012)
}
