% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FISK.R
\name{se_shape_FISK}
\alias{se_shape_FISK}
\title{Standard Error of Fisk Shape Parameter}
\usage{
se_shape_FISK(Gini, se_Gini)
}
\arguments{
\item{Gini}{Numeric in (0,1). The Gini coefficient}

\item{se_Gini}{Standard error of the Gini estimate.}
}
\value{
Standard error of the Fisk shape parameter.
}
\description{
Uses the delta method: \eqn{shape = 1/Gini}.
}
\examples{
se_shape_FISK (0.36, 0.012)
}
