% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FISK.R
\name{shape_Fisk}
\alias{shape_Fisk}
\title{Mean-Matching Shape Parameter for Fisk Distribution}
\usage{
shape_Fisk(Gini)
}
\arguments{
\item{Gini}{Numeric in (0,1). The Gini coefficient}
}
\value{
Shape parameter.
}
\description{
Mean-Matching Shape Parameter for Fisk Distribution
}
\examples{
shape_Fisk (0.367)
}
